fixed bug in parser.cpp for the DROP TABLE command that would make it so that the code works with only "DROP (table name)", works properly now.

+++ Personal Database/Parser.cpp
@@ -209,7 +209,20 @@ std::unique_ptr<SQLCommand> parser::parseDropTable()
 {
 	auto sqlcmd = std::make_unique<dropTable>();
 
-	if (!tokenHead.hasMoreTokens()) {
+	if (!tokenHead.hasMoreTokens())
+	{
+		throw std::runtime_error("Expected TABLE keyword after DROP");
+	}
+
+	token nextToken = tokenHead.getNextToken();
+
+	if (nextToken.text != "TABLE")
+	{
+		throw std::runtime_error("Expected TABLE keyword after DROP");
+	}
+
+	if (!tokenHead.hasMoreTokens())
+	{
 		throw std::runtime_error("Expected table name after DROP TABLE");
 	}
